skypulse-telegram/bot.py
"""
SkyPulse Telegram Bot
AI-powered flight deal notification bot
"""
import os
import logging
from datetime import datetime
from telegram import Update, InlineKeyboardButton, InlineKeyboardMarkup
from telegram.ext import Application, CommandHandler, MessageHandler, CallbackQueryHandler, ContextTypes, filters

# Configuration
TELEGRAM_BOT_TOKEN = os.getenv('TELEGRAM_BOT_TOKEN', 'YOUR_BOT_TOKEN')
DATABASE_URL = os.getenv('DATABASE_URL', 'sqlite:///skypulse.db')

# Enable logging
logging.basicConfig(
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    level=logging.INFO
)
logger = logging.getLogger(__name__)

# User sessions
user_sessions = {}

async def start_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    \"\"\"Handle /start command\"\"\"
    user = update.effective_user
    welcome_text = f\"\"\"
✈️ Welcome to SkyPulse!

I'm your AI-powered flight deal assistant.

Features:
• Get notified about cheap flights
• Price predictions & trends
• Personalized recommendations

Commands:
/start - Show this message
/subscribe - Subscribe to flight alerts
/unsubscribe - Unsubscribe from alerts
/deals - View latest deals
/help - Show help

Get started: /subscribe
\"\"\"
    await update.message.reply_text(welcome_text)

async def help_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    \"\"\"Handle /help command\"\"\"
    help_text = \"\"\"
📖 SkyPulse Help

Commands:
/start - Start the bot
/subscribe - Subscribe to flight alerts
/unsubscribe - Unsubscribe
/deals - View latest deals
/alerts - Manage your alerts
/profile - View your profile

How it works:
1. Subscribe to alerts
2. Tell me your travel preferences
3. I'll notify you when deals match!

Example: "Notify me about flights to Paris under 500EUR"
\"\"\"
    await update.message.reply_text(help_text)

async def subscribe_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    \"\"\"Handle /subscribe command\"\"\"
    keyboard = [
        [
            InlineKeyboardButton("✈️ Flight Deals", callback_data="subscribe_flights"),
            InlineKeyboardButton("🏨 Hotel Deals", callback_data="subscribe_hotels")
        ],
        [InlineKeyboardButton("✅ Confirm", callback_data="confirm_subscribe")]
    ]
    reply_markup = InlineKeyboardMarkup(keyboard)
    
    await update.message.reply_text(
        "Select what you want to be notified about:",
        reply_markup=reply_markup
    )

async def deals_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    \"\"\"Handle /deals command - show latest deals\"\"\"
    # This would fetch from database in real implementation
    sample_deals = [
        {
            "route": "Berlin → Paris",
            "price": "€89",
            "date": "2026-03-15",
            "airline": "EasyJet"
        },
        {
            "route": "Munich → London",
            "price": "€120",
            "date": "2026-03-20",
            "airline": "British Airways"
        },
        {
            "route": "Frankfurt → NYC",
            "price": "€450",
            "date": "2026-04-01",
            "airline": "Lufthansa"
        }
    ]
    
    deals_text = "✈️ Latest Flight Deals:\n\n"
    for i, deal in enumerate(sample_deals, 1):
        deals_text += f"{i}. {deal['route']}\n"
        deals_text += f"   💰 {deal['price']} | 📅 {deal['date']} | ✈️ {deal['airline']}\n\n"
    
    keyboard = [
        [InlineKeyboardButton("🔔 Subscribe to Alerts", callback_data="subscribe_deals")]
    ]
    reply_markup = InlineKeyboardMarkup(keyboard)
    
    await update.message.reply_text(deals_text, reply_markup=reply_markup)

async def handle_message(update: Update, context: ContextTypes.DEFAULT_TYPE):
    \"\"\"Handle user messages - natural language processing\"\"\"
    user_text = update.message.text
    
    # Simple NLP - detect intent
    user_text_lower = user_text.lower()
    
    if any(word in user_text_lower for word in ['paris', 'berlin', 'munich', 'london', 'new york']):
        response = "I'd love to help you find flights! Use /subscribe to get alerts for your preferred destination."
    elif any(word in user_text_lower for word in ['cheap', 'deal', 'discount', 'offer']):
        response = "Great! Check out the latest deals with /deals command!"
    elif any(word in user_text_lower for word in ['subscribe', 'alert', 'notify']):
        response = "Use /subscribe to set up flight alerts!"
    else:
        response = "I can help you find flight deals! Try:\n• /deals - View latest deals\n• /subscribe - Get notified\n• Tell me a destination (e.g., 'flights to Paris')"
    
    await update.message.reply_text(response)

async def button_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
    \"\"\"Handle inline button callbacks\"\"\"
    query = update.callback_query
    await query.answer()
    
    if query.data == "subscribe_flights":
        await query.edit_message_text(
            "✅ You're now subscribed to flight deal alerts!\n\n"
            "You'll be notified when we find great deals matching your preferences."
        )
    elif query.data == "confirm_subscribe":
        await query.edit_message_text(
            "🎉 Subscription confirmed!\n\n"
            "Use /deals to see current offers or /alerts to customize your preferences."
        )

async def error_handler(update: Update, context: ContextTypes.DEFAULT_TYPE):
    \"\"\"Handle errors\"\"\"
    logger.error(f"Update {update} caused error {context.error}")

def main():
    \"\"\"Run the bot\"\"\"
    application = Application.builder().token(TELEGRAM_BOT_TOKEN).build()
    
    # Add handlers
    application.add_handler(CommandHandler("start", start_command))
    application.add_handler(CommandHandler("help", help_command))
    application.add_handler(CommandHandler("subscribe", subscribe_command))
    application.add_handler(CommandHandler("deals", deals_command))
    application.add_handler(MessageHandler(filters.TEXT & ~filters.COMMAND, handle_message))
    application.add_handler(CallbackQueryHandler(button_callback))
    
    # Error handler
    application.add_error_handler(error_handler)
    
    # Start polling
    logger.info("SkyPulse Bot starting...")
    application.run_polling(allowed_updates=Update.ALL_TYPES)

if __name__ == "__main__":
    main()
