skypulse-telegram/requirements.txt
python-telegram-bot==20.6
python-dotenv>=1.0.0
requests>=2.31.0
